requirements.txt
flask==3.0.0
claude-agent-sdk==0.1.0
pandas==2.1.4
openpyxl==3.1.2
plotly==5.18.0
kaleido==0.2.1
python-dotenv==1.0.0
bcrypt==4.1.2
PyYAML==6.0.1
psycopg2-binary==2.9.9
